Correcting abstract
Commit related with issue #1 and #6, but the motivation about proposed termination conditions and metrics is still needed (discuss with @amorag and @deantares), so I can't close these issues yet.
Corrected some typos, long sentences and repeated words.
The abstract is coherent with the stuff we want to do, so it is OK for me to keep it.